removing player config
