+++ speed-dreams/source-2.2.3/src/modules/userinterface/legacymenu/CMakeLists.txt
@@ -14,7 +14,7 @@ IF(MSVC)
 ENDIF(MSVC)
 
 SET(CS confscreens)
-SET(CONFSCREENS_SOURCES ${CS}/playerconfig.cpp ${CS}/controlconfig.cpp
+SET(CONFSCREENS_SOURCES ${CS}/controlconfig.cpp
                         ${CS}/joystickconfig.cpp ${CS}/mouseconfig.cpp
                         ${CS}/joy2butconfig.cpp
                         ${CS}/displayconfig.cpp ${CS}/graphconfig.cpp ${CS}/openglconfig.cpp
@@ -22,7 +22,7 @@ SET(CONFSCREENS_SOURCES ${CS}/playerconfig.cpp ${CS}/controlconfig.cpp
                         ${CS}/simuconfig.cpp ${CS}/soundconfig.cpp ${CS}/aiconfig.cpp
                         ${CS}/hostsettingsmenu.cpp)
 SET(CONFSCREENS_HEADERS ${CS}/confscreens.h ${CS}/advancedgraphconfig.h
-                        ${CS}/playerconfig.h ${CS}/controlconfig.h
+                        ${CS}/controlconfig.h
                         ${CS}/joystickconfig.h ${CS}/mouseconfig.h
                         ${CS}/joy2butconfig.h ${CS}/monitorconfig.h
                         ${CS}/displayconfig.h ${CS}/graphconfig.h ${CS}/openglconfig.h
