--- speed-dreams/source-2.2.3/src/modules/userinterface/legacymenu/confscreens/playerconfig.h
@@ -1,28 +1,0 @@
-/***************************************************************************
-
-    file                 : playerconfig.h
-    created              : Wed Apr 26 22:20:10 CEST 2000
-    copyright            : (C) 2000 by Eric Espie
-    email                : torcs@free.fr
-    version              : $Id: playerconfig.h 3438 2011-03-11 13:43:24Z pouillot $
-
- ***************************************************************************/
-
-/***************************************************************************
- *                                                                         *
- *   This program is free software; you can redistribute it and/or modify  *
- *   it under the terms of the GNU General Public License as published by  *
- *   the Free Software Foundation; either version 2 of the License, or     *
- *   (at your option) any later version.                                   *
- *                                                                         *
- ***************************************************************************/
-
-#ifndef _PLAYERCONFIG_H_
-#define _PLAYERCONFIG_H_
-
-#include "confscreens.h"
-
-
-extern void *PlayerConfigMenuInit(void *prevMenu);
-
-#endif /* _PLAYERCONFIG_H_ */
