+++ speed-dreams/source-2.2.3/src/modules/userinterface/legacymenu/racescreens/networkingmenu.cpp
@@ -41,7 +41,7 @@
 #include <drivers.h>
 #include <network.h>
 
-#include <playerconfig.h>
+#include <controlconfig.h>
 #include <playerpref.h>
 #include <hostsettingsmenu.h>
