+++ speed-dreams/source-2.2.3/src/modules/userinterface/legacymenu/racescreens/racestopmenu.cpp
@@ -20,7 +20,6 @@
 
 #include <isoundengine.h>
 
-#include <playerconfig.h>
 #include <controlconfig.h>
 
 #include "legacymenu.h"
